test: Reinstate text highlight without selection
Our `typeInRichText` integration helper does not support parsing inner
HTML tags. This is required in order to type at the correct cursor
position, i.e. with the cursor placed within the `mark` tag that applies
the text highlighting.

Adding support for this gets complicated quickly, e.g. how might we
support typing with a text selection the begins within a tag but
finishes outside of it?

Instead, adding an e2e test is likely the best approach.